#166 Refactor: ExchangeRequestDto -> ExchangeAddRequestDto 이름 변경 및 내용 수정
- JavaDoc 내용 추가
- 내장되어있던 Dto 변환 Method 삭제
